Fixed failing config upgrade from pre-7.4.0 .pstate files.

+++ Tribler/Core/Upgrade/config_converter.py
@@ -5,6 +5,7 @@ import base64
 import logging
 import os
 from glob import iglob
+from lib2to3.pgen2.parse import ParseError
 
 from configobj import ConfigObj
 
@@ -205,44 +206,64 @@ def convert_config_to_tribler74(state_dir=None):
     """
     Convert the download config files to Tribler 7.4 format. The extensions will also be renamed from .state to .conf
     """
+    refactoring_tool = None
     if PY3:
         from lib2to3.refactor import RefactoringTool, get_fixers_from_package
         refactoring_tool = RefactoringTool(fixer_names=get_fixers_from_package('lib2to3.fixes'))
 
     state_dir = state_dir or TriblerConfig.get_default_base_state_dir()
     for _, filename in enumerate(iglob(os.path.join(state_dir, STATEDIR_CHECKPOINT_DIR, '*.state'))):
-        old_config = CallbackConfigParser()
-        try:
-            old_config.read_file(filename)
-        except MissingSectionHeaderError:
-            logger.error("Removing download state file %s since it appears to be corrupt", filename)
-            os.remove(filename)
+        convert_state_file_to_conf_74(filename, refactoring_tool=refactoring_tool)
 
-        # We first need to fix the .state file such that it has the correct metainfo/resumedata
+
+def convert_state_file_to_conf_74(filename, refactoring_tool=None):
+    """
+    Converts .pstate file (pre-7.4.0) to .conf file.
+    :param filename: .pstate file
+    :param refactoring_tool: RefactoringTool instance if using Python3
+    :return: None
+    """
+    def _fix_state_config(config):
         for section, option in [('state', 'metainfo'), ('state', 'engineresumedata')]:
-            value = old_config.get(section, option, literal_eval=False)
+            value = config.get(section, option, literal_eval=False)
             ungarbled_dict = None
-            if PY3:
-                value = str(refactoring_tool.refactor_string(value+'\n', option + '_2to3'))
-                ungarbled_dict = recursive_ungarble_metainfo(ast.literal_eval(value))
             try:
+                if PY3 and refactoring_tool:
+                    value = str(refactoring_tool.refactor_string(value + '\n', option + '_2to3'))
+                    ungarbled_dict = recursive_ungarble_metainfo(ast.literal_eval(value))
+
                 value = ungarbled_dict or ast.literal_eval(value)
-                old_config.set(section, option, base64.b64encode(lt.bencode(value)).decode('utf-8'))
-            except (ValueError, SyntaxError):
-                logger.error("Removing download state file %s since it could not be converted", filename)
-                os.remove(filename)
-                continue
+                config.set(section, option, base64.b64encode(lt.bencode(value)).decode('utf-8'))
+            except (ValueError, SyntaxError, ParseError) as ex:
+                logger.error("Config could not be fixed, probably corrupted. Exception: %s %s", type(ex), str(ex))
+                return None
+        return config
 
-        # Remove dlstate since the same information is already stored in the resumedata
-        if old_config.has_option('state', 'dlstate'):
-            old_config.remove_option('state', 'dlstate')
-
-        new_config = ConfigObj(infile=filename[:-6] + '.conf', encoding='utf8')
-        for section in old_config.sections():
-            for key, _ in old_config.items(section):
-                val = old_config.get(section, key)
-                if section not in new_config:
-                    new_config[section] = {}
-                new_config[section][key] = val
-        new_config.write()
+    old_config = CallbackConfigParser()
+    try:
+        old_config.read_file(filename)
+    except MissingSectionHeaderError:
+        logger.error("Removing download state file %s since it appears to be corrupt", filename)
         os.remove(filename)
+
+    # We first need to fix the .state file such that it has the correct metainfo/resumedata.
+    # If the config cannot be fixed, it is likely corrupted in which case we simply remove the file.
+    fixed_config = _fix_state_config(old_config)
+    if not fixed_config:
+        logger.error("Removing download state file %s since it could not be fixed", filename)
+        os.remove(filename)
+        return
+
+    # Remove dlstate since the same information is already stored in the resumedata
+    if fixed_config.has_option('state', 'dlstate'):
+        fixed_config.remove_option('state', 'dlstate')
+
+    new_config = ConfigObj(infile=filename[:-6] + '.conf', encoding='utf8')
+    for section in fixed_config.sections():
+        for key, _ in fixed_config.items(section):
+            val = fixed_config.get(section, key)
+            if section not in new_config:
+                new_config[section] = {}
+            new_config[section][key] = val
+    new_config.write()
+    os.remove(filename)
